implemented Doctors stage

+++ src/ru/clinic/application/java/fx/controllers/ControllerDbTables.java
@@ -11,7 +11,9 @@ import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import ru.clinic.application.java.fx.frames.FrameDbTables;
+import ru.clinic.application.java.service.AdminService;
 import ru.clinic.application.java.service.dataBaseModel.TableAdmins;
+import ru.clinic.application.java.service.dataBaseModel.TableDoctors;
 
 import java.util.Optional;
 
@@ -33,11 +35,21 @@ public class ControllerDbTables {
     private final static String CONFIRMATION_HEADER = "Вы собираетесь изменить структуру базы данных. \nЕсли удаляется таблица, то это приведет к потери всех данных";
     private final static String CONFIRMATION_CONTENT = "Вы уверены, что хотите продолжить?";
 
+    /*Information Dialog not authorised*/
+    private static final String INFORMATION_TITLE = "Информационное окно";
+    private static final String INFORMATION_CONTEXT = "У Вас не хватает прав для выполнения данной операции. \nЗайдите под главным администратором.";
+
     @Autowired
     TableAdmins tableAdmins;
 
     @Autowired
+    TableDoctors tableDoctors;
+
+    @Autowired
     FrameDbTables frameDbTables;
+
+    @Autowired
+    AdminService adminService;
 
     @FXML
     private Label adminStatus;
@@ -54,17 +66,33 @@ public class ControllerDbTables {
     @FXML
     private Button closeFrameBtn;
 
+    @FXML
+    private Label doctorStatus;
+
+    @FXML
+    private Button createDoctor;
+
+    @FXML
+    private Button dropDoctor;
+
+    @FXML
+    private Button updateDoctor;
+
 
 
 
     public void startController() {
         setAdminStatus();
+        setDoctorsStatus();
     }
 
-    public void setAdminStatus() {
-        if (adminStatus != null && createAdmin != null && dropAdmin != null){
-            setCreatedStatus(tableAdmins.isCreated(), adminStatus, createAdmin, dropAdmin);
-        }
+    private void alertNoAuth() {
+        Alert alert = new Alert(Alert.AlertType.INFORMATION);
+        alert.setTitle(INFORMATION_TITLE);
+        alert.setHeaderText(null);
+        alert.setContentText(INFORMATION_CONTEXT);
+
+        alert.showAndWait();
     }
 
     private void setCreatedStatus(boolean created, Label labelStatus, Button createBtn, Button dropBtn){
@@ -89,11 +117,20 @@ public class ControllerDbTables {
 
         Optional<ButtonType> result = alert.showAndWait();
         if (result.get() == ButtonType.OK){
-            System.out.println("True");
             return true;
         }
-        System.out.println("False");
         return false;
+    }
+
+    @FXML
+    void closeFrameAction(ActionEvent event) {
+        frameDbTables.stop();
+    }
+
+    public void setAdminStatus() {
+        if (adminStatus != null && createAdmin != null && dropAdmin != null){
+            setCreatedStatus(tableAdmins.isCreated(), adminStatus, createAdmin, dropAdmin);
+        }
     }
 
     @FXML
@@ -105,8 +142,12 @@ public class ControllerDbTables {
 
     @FXML
     void dropAdminAction(ActionEvent event) {
-        if (confirmAction()) {
-            tableAdmins.dropTable();
+        if (adminService.getCurrentAdmin().getId() == 0) {
+            if (confirmAction()) {
+                tableAdmins.dropTable();
+            }
+        }else {
+            alertNoAuth();
         }
     }
 
@@ -115,8 +156,32 @@ public class ControllerDbTables {
         tableAdmins.checkIfCreated();
     }
 
+    public void setDoctorsStatus() {
+        if (doctorStatus != null && createDoctor != null && dropDoctor != null){
+            setCreatedStatus(tableDoctors.isCreated(), doctorStatus, createDoctor, dropDoctor);
+        }
+    }
+
     @FXML
-    void closeFrameAction(ActionEvent event) {
-        frameDbTables.stop();
+    void createDoctorAction(ActionEvent event) {
+        if (confirmAction()) {
+            tableDoctors.createTable();
+        }
+    }
+
+    @FXML
+    void dropDoctorAction(ActionEvent event) {
+        if (adminService.getCurrentAdmin().getId() == 0) {
+            if (confirmAction()) {
+                tableDoctors.dropTable();
+            }
+        }else {
+            alertNoAuth();
+        }
+    }
+
+    @FXML
+    void updateDoctorAction(ActionEvent event) {
+        tableDoctors.checkIfCreated();
     }
 }
